pyproject.toml
[project]
name = "adk-playgound"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.31.1",
    "jupyter>=1.1.1",
    "litellm>=1.83.14",
]
